+++ files/default/cloudwatch_logs/write_cloudwatch_agent_json.py
@@ -37,7 +37,7 @@ def parse_args():
                              '(i.e., is it a compute node or the head node?)')
     parser.add_argument('--scheduler',
                         required=True,
-                        choices=['slurm', 'sge', 'torque', 'awsbatch'],
+                        choices=['slurm', 'awsbatch'],
                         help='Scheduler')
     return parser.parse_args()